Rename HandshakeRequester to MessageRequest
