+++ TheAgencyEJB/ejbModule/service/MessageRequestLocal.java
@@ -1,4 +1,4 @@
-package handshake;
+package service;
 
 import java.io.IOException;
 import java.util.concurrent.TimeoutException;
@@ -9,7 +9,7 @@ import exceptions.ConnectionException;
 import model.ServiceMessage;
 
 @Local
-public interface HandshakeRequesterLocal {
+public interface MessageRequestLocal {
 	
 	public ServiceMessage sendMessage(String destination, ServiceMessage message) throws ConnectionException, IOException, TimeoutException, InterruptedException;
