+++ TheAgencyEJB/ejbModule/service/MessageResponseLocal.java
@@ -1,11 +1,11 @@
-package handshake;
+package service;
 
 import java.io.IOException;
 
 import javax.ejb.Local;
 
 @Local
-public interface HandshakeResponseLocal {
+public interface MessageResponseLocal {
 
 	public void waitMessage() throws IOException;
 }
